feat: Add ability to generate changelogs across Ubuntu release and also better handling of binary package version to source package version lookups
This is a breaking changes as it requires two new args `--from-series` and `--to-series` so you can inform the tool what the
Ubuntu series/release was for the "from" manifest and the "to" manifest.

For example `--from-series bionic --to-series focal` means the "from" manifest is from a bionic image and the "to" manifest is from a focal image.

Bug fixes around binary package names not being the same as source package names also included. This was solved by doing a lookup to the Launchpad
API to find the corresponding source package name and source package version for each binary package name and binary package version.

This can lead to duplicated changelogs for binary packages with the same source package but it adds context so these are not removed.