feat: Add support for including CVE details from each changelog and outputting to JSON file
One of the main use cases for creating a changelog between manifests is to be able to track the CVEs addressed.

This commit adds the ability to highlight these CVEs using the new `--highlight-cves` flag and also to include the
description of those CVEs using the `--highlight-cves-show-cve-description` flag.

One feature previously lacking was JSON output or an output that can be parsed in automation.

JSON was the obvious choice. Using JSON output `--output-json changelog.json` in combination with `--highlight-cves`
will add even more detail about all CVEs, example:

```
"cves": [
        {
            "cve": "CVE-2022-32221",
            "url": "https://ubuntu.com/security/CVE-2022-32221",
            "cve_description": " [POST following PUT confusion]",
            "cve_priority": "medium",
            "cve_public_date": "2022-10-26 07:00:00 UTC"
        },
        {
            "cve": "CVE-2022-35252",
            "url": "https://ubuntu.com/security/CVE-2022-35252",
            "cve_description": " When curl is used to retrieve and parse cookies from a HTTP(S) server, itaccepts cookies using control codes that when later are sent back to a HTTPserver might make the server return 400 responses. Effectively allowing a\"sister site\" to deny service to all siblings.",
            "cve_priority": "low",
            "cve_public_date": "2022-09-23 14:15:00 UTC"
        }
    ],
```