fix: Improve buggy changelog diff logic
Currently, the tool determines changes from one package version to the
another by computing a unified diff of the changelogs.  However, this
logic fails when a few change-describing lines are moved from under an
older version to the change block under a newer version. A unified diff
between such versions omits the lines carried forward.  This commit
introduces changes that improve this logic.  With this commit, when
determining the changes from one package version to another, the tool
now finds all version entries in the changelog that are present in the
newer package changelog but absent from the older package changelog. The
change block entries under all the determined versions are then
processed.